+++ gorang/src/main/java/com/kh/gorang/member/controller/MemberController.java
@@ -8,23 +8,25 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 public class MemberController {
 	
+	//마이페이지 메인
 	@RequestMapping("myPageAllView.me")
 	public String myPageViewAll(){
 		return "member/myPageAllView";
 	}
 	
-	
+	//마이페이지 레시피게시판
 	@RequestMapping("myRecipeBoard.me")
 	public String myRecipeBoard(){
 		return "member/myRecipeBoard";
 	}
 	
-
+	//마이페이지 나의냉장고
 	@RequestMapping("myRefrigerator.me")
 	public String myRefrigerator(){
 		return "member/myRefrigerator";
 	}
 	
+	//내 정보 수정
 	@RequestMapping("myPageInfoEdit.me")
 	public String myPageInfoEdit(){
 		return "member/myPageInfoEdit";
@@ -35,7 +37,19 @@ public class MemberController {
 	public String myPageBuyList(){
 		return "member/myPageBuyList";
 	}
-
+	
+	//마이 페이지 질의응답
+	@RequestMapping("myPageQnA.me")
+	public String myPageQnAView(){
+		return "member/myPageQnA";
+	}
+	
+	//마이 페이지 리뷰
+	@RequestMapping("myPageReplyReview.me")
+	public String myPageReplyReviewView(){
+		return "member/myPageReplyReview";
+	}
+	
 }
