--- gorang/src/main/java/com/kh/gorang/member/controller/MemberQnAController.java
@@ -1,18 +1,0 @@
-package com.kh.gorang.member.controller;
-
-//import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-//import org.springframework.web.servlet.ModelAndView;
-
-@Controller
-public class MemberQnAController {
-	
-	@RequestMapping("myPageQnA.me")
-	public String myPageQnAView(){
-		return "member/myPageQnA";
-	}
-	
-}
-
-
